many float/int division/printing bugs being squashed

+++ codebase/workers/unity/Assets/Gamelogic/Controller/ControllerBehaviour.cs
@@ -46,7 +46,7 @@ public class ControllerBehaviour : MonoBehaviour
 
 	int usedSlots;
 
-	int revenue;
+	float revenue;
 	float costs;
 	float penalties;
